Refatora serviços de aplicação em módulos dedicados
